+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed `ChangeInterval` rule to `ChangeWithIntervalRule`.
 - Renamed `CharacterClass` rule to `CharacterClassRule`.
 - Renamed `Dictionary` rule to `DictionaryRule`.
+- Renamed `Digit` rule to `DigitRule`.
 
 ### Fixed
 - Nothing